backend/requirements.txt
# Django и основные зависимости
python-dotenv==1.0.0
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Аутентификация и безопасность
djangorestframework-simplejwt==5.3.0
django-encrypted-model-fields==0.5.8

# Асинхронность и фоновые задачи
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Работа с файлами и медиа
Pillow==10.1.0
python-magic==0.4.27

# Утилиты для работы с данными
djangorestframework-csv==2.1.1
openpyxl==3.1.2
PyPDF2==3.0.1
pandas==2.1.3

# API и сериализация
drf-yasg==1.21.7
markdown==3.5.1
uritemplate==4.1.1

# Обработка электронной почты
exchangelib==4.9.0
beautifulsoup4==4.12.2

# WhatsApp интеграция
selenium==4.15.2

# Генерация и обработка документов
reportlab==4.0.7
docxtpl==0.16.4

# Валидация и обработка
jsonschema==4.19.2
phonenumbers==8.13.24
python-slugify==8.0.1

# Безопасность и криптография
cryptography==41.0.5
social-auth-app-django==5.4.0

# Кэширование
django-redis==5.4.0

# Логирование
sentry-sdk==1.38.0

# Работа с изображениями
opencv-python-headless==4.8.1.78

# Математические и научные вычисления
numpy==1.26.2
scipy==1.11.4

# Дополнительные утилиты
python-dateutil==2.8.2
pytz==2023.3.post1

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1

# Разработка и отладка
ipython==8.17.2
black==23.11.0
flake8==6.1.0
isort==5.12.0